tg orders in progress

+++ server/src/services/user.service.js
@@ -54,10 +54,9 @@ class UserService {
     async deleteById(id) {
         try {
             const userToDelete = await User.findById(id);
-            // TODO В майбутньому треба буде розкоментувати
-            // if (userToDelete.role === EUserRole.ADMIN) {
-            //     throw new ApiError("You can not delete another admin", 400)
-            // }
+            if (userToDelete.role === EUserRole.ADMIN) {
+                throw new ApiError("You can not delete another admin", 400)
+            }
 
             await Promise.all([
                 Action.deleteMany({_userId: id}),
